Feature #177: add loadingScreen
